+++ pymongo/mongo_client.py
@@ -350,8 +350,8 @@ class MongoClient(common.BaseObject):
                            unicode(password), mechanism)
             try:
                 self._cache_credentials(source, credentials)
-            except OperationFailure:
-                raise ConfigurationError("authentication failed")
+            except OperationFailure, exc:
+                raise ConfigurationError(str(exc))
 
     def _cached(self, dbname, coll, index):
         """Test if `index` is cached.